maj media queries & fix bugs

+++ projet_4_Forteroche/src/view/adminSingleCom.php
@@ -18,8 +18,17 @@
         </p>
 
         <div class="actions">
-            <button type="button" name="update" class="button1"><a href="index.php?action=commentUpdate&amp;id=<?= $Comment->id(); ?>">Editer <i class="fas fa-pen-nib"></i></a></button>
-            <button type="button" name="delete" class="button3"> <a href="index.php?action=deleteCom&amp;id=<?= $Comment->id()?>">Supprimer <i class="fas fa-trash-alt"></i></a></button>
+            <?php if($Comment->report() == 1 && $_SESSION['user_role'] == 2): ?>
+                <button type="button" name="validate" class="button2"> <a href="index.php?action=validateCom&amp;id=<?= $Comment->id(); ?>">Valider <i class="fas fa-check-circle"></i></a></button>
+            <?php endif; ?>
+            
+            <?php if($Comment->pseudo() == $_SESSION['pseudo']): ?>
+                <button type="button" name="update" class="button1"><a href="index.php?action=commentUpdate&amp;id=<?= $Comment->id(); ?>">Editer <i class="fas fa-pen-nib"></i></a></button>
+            <?php endif; ?>
+
+            <?php if($Comment->user_id() == $_SESSION['id'] || $_SESSION['user_role'] == 2): ?>
+                <button type="button" name="delete" class="button3"> <a href="index.php?action=deleteCom&amp;id=<?= $Comment->id()?>">Supprimer <i class="fas fa-trash-alt"></i></a></button>
+            <?php endif; ?>
         </div>
 
     </article>
